script: Move document DOM interfaces to script/dom/document/
