+++ components/script/dom/document/documentorshadowroot.rs
@@ -21,7 +21,6 @@ use style::stylesheets::{Stylesheet, StylesheetContents};
 use stylo_atoms::Atom;
 use webrender_api::units::LayoutPoint;
 
-use super::bindings::trace::HashMapTracedValues;
 use crate::dom::bindings::cell::DomRefCell;
 use crate::dom::bindings::codegen::Bindings::NodeBinding::GetRootNodeOptions;
 use crate::dom::bindings::codegen::Bindings::NodeBinding::Node_Binding::NodeMethods;
@@ -30,6 +29,7 @@ use crate::dom::bindings::conversions::{ConversionResult, SafeFromJSValConvertib
 use crate::dom::bindings::inheritance::Castable;
 use crate::dom::bindings::num::Finite;
 use crate::dom::bindings::root::{Dom, DomRoot};
+use crate::dom::bindings::trace::HashMapTracedValues;
 use crate::dom::css::stylesheetlist::StyleSheetListOwner;
 use crate::dom::element::Element;
 use crate::dom::html::htmlelement::HTMLElement;
